Conflicts:
	applications/plugins/Compliant/Compliant_test/CompliantImplicitSolver_test.cpp
	applications/plugins/Compliant/Compliant_test/Compliant_test.h
	applications/plugins/Compliant/Compliant_test/DampedOscillator_test.cpp
	applications/plugins/Flexible/Flexible_test/HexahedraMaterial_test.cpp
	applications/plugins/Flexible/Flexible_test/TetrahedraMaterial_test.cpp
	modules/sofa/simulation/common/MechanicalVisitor.h

Former-commit-id: d7bbae7780ca85be8a4c4d29e864f69e38a0dbf0

+++ modules/SofaMeshCollision/IncrSAP.inl
@@ -383,7 +383,7 @@ void TIncrSAP<List,Allocator>::showEndPoints()const{
 
 template <template<class T,class Allocator> class List,template <class T> class Allocator>
 void TIncrSAP<List,Allocator>::showBoxes()const{
-    for(int i = 0 ; i < _boxes.size() ; ++i){
+    for(size_t i = 0 ; i < _boxes.size() ; ++i){
         const ISAPBox & box = _boxes[i];
         std::cout<<"collision model "<<box.cube.getCollisionModel()->getLast()<<" index "<<box.cube.getExternalChildren().first.getIndex()<<std::endl;
